Enable deletion for the selfUser for failed to send messages and in left conversations (#397)
* Enable deletion for the selfUser for failed to send messages and in left conversations

* Update for new specifications

* If a message is failed to send only "delete for everyone" is possible as it might have it the backend before we expired it.
* If a message is sent by self and not expired it can be deleted for send or everyone.
* Ephemeral messages can not be deleted (no menu at all).
* If the user leaves a conversation he can delete previously sent messages only using "delete for me".

+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConversationCell.m
@@ -541,8 +541,11 @@ const NSTimeInterval ConversationCellSelectionAnimationDuration = 0.33;
         NSString *likeTitleKey = [Message isLikedMessage:self.message] ? @"content.message.unlike" : @"content.message.like";
         UIMenuItem *likeItem = [[UIMenuItem alloc] initWithTitle:NSLocalizedString(likeTitleKey, @"") action:@selector(likeMessage:)];
         [items insertObject:likeItem atIndex:0];
+    }
 
-        UIMenuItem *deleteItem = [[UIMenuItem alloc] initWithTitle:NSLocalizedString(@"content.message.delete", @"") action:@selector(deleteMessage:)];
+    if (self.message.canBeDeleted) {
+        NSString *deleteTitle = NSLocalizedString(@"content.message.delete", @"");
+        UIMenuItem *deleteItem = [[UIMenuItem alloc] initWithTitle:deleteTitle action:@selector(deleteMessage:)];
         [items addObject:deleteItem];
     }
